refactor echo and add component ID(5015)
