+++ plugins/echo/go.mod
@@ -1,17 +1,14 @@
-module github.com/apache/skywalking-go/plugins/echov4
+module github.com/apache/skywalking-go/plugins/echo
 
 go 1.20
 
-require (
-	github.com/apache/skywalking-go/plugins/core v0.0.0-20230712143836-5d2a80f0e1f5
-	github.com/labstack/echo/v4 v4.11.1
-)
+require github.com/labstack/echo/v4 v4.11.1
 
 require (
-	github.com/dave/dst v0.27.2 // indirect
 	github.com/labstack/gommon v0.4.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/stretchr/testify v1.8.2 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasttemplate v1.2.2 // indirect
 	golang.org/x/crypto v0.11.0 // indirect
