+++ test/plugins/scenarios/echo/bin/startup.sh
