+++ test/plugins/scenarios/echo/config/excepted.yml
@@ -26,7 +26,7 @@ segmentItems:
             spanLayer: Http
             startTime: nq 0
             endTime: nq 0
-            componentId: 5006
+            componentId: 5015
             isError: false
             spanType: Entry
             peer: ""
@@ -69,7 +69,7 @@ segmentItems:
             spanLayer: Http
             startTime: gt 0
             endTime: gt 0
-            componentId: 5006
+            componentId: 5015
             isError: false
             spanType: Entry
             peer: ""
